(styling) move the unstyled button in its own helper file
we will certainly want to create unstyled `<a>`, `<ul>`, etc. Just keep
those in the same file
